Move model rendering code to numpyro.contrib.render
