+++ numpyro/__init__.py
@@ -5,7 +5,8 @@ from numpyro import compat, diagnostics, distributions, handlers, infer, optim
 from numpyro.distributions.distribution import enable_validation, validation_enabled
 import numpyro.patch  # noqa: F401
 from numpyro.primitives import deterministic, factor, module, param, plate, plate_stack, prng_key, sample, subsample
-from numpyro.util import enable_x64, set_host_device_count, set_platform, render_model
+from numpyro.util import enable_x64, set_host_device_count, set_platform
+from numpyro.contrib.render import render_model
 from numpyro.version import __version__
 
 set_platform('cpu')
